my-react-app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
meta==1.0.2
platformdirs==2.5.2
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.34
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
zipp==3.8.0
